functions/chunkify/requirements.txt
google-cloud-storage
google-cloud-firestore
google-cloud-pubsub